### 2019-02-23 #1274 @arfon
Cross-linking with rOpenSci here: https://github.com/ropensci/software-review/issues/269
