### 2019-02-23 #1276 @whedon
[REVIEW]: Ecological Metadata as Linked Data
**Submitting author:** @cboettig (<a href="http://orcid.org/0000-0002-1642-628X">Carl Boettiger</a>)
**Repository:** <a href="https://github.com/ropensci/emld" target ="_blank">https://github.com/ropensci/emld</a>
**Version:** 0.2.0
