### 2019-02-24 #1276 @cboettig
@arfon @usethedata just a heads up that this software was already just peer reviewed under rOpenSci with @lmullen as editor, see https://github.com/ropensci/software-review/issues/269
