### 2019-02-24 #1276 @usethedata
Completed expedited review given rOpenSci peer review at /ropensci/software-review/issues/269.  Manuscript generated properly.
