### 2019-02-24 #1276 @arfon
@usethedata - if this looks good, the next step is to assign the archive for the submission. We should also check the references too... @cboettig - is there already an archive of this repository?
