### 2019-02-24 #1276 @arfon
Looks like this monstrosity of a DOI is causing @whedon some problems: https://github.com/ropensci/emld/blob/master/paper.bib#L93
