### 2019-02-24 #1276 @usethedata
@arfon @cboettig I did check that all of the links in the references resolved to the paper or web site that was referenced.  Looking a bit more closely, one journal title is abbreviated (Michener et al 1997, 5th from the bottom of the manuscript).  That's a nit I'd fix if it were my paper.
