### 2019-02-25 abeed8d
add CRAN comments
