+++ cran-comments.md
@@ -1,0 +1,19 @@
+Dear CRAN,
+
+I am pleased to submit the initial release of the emld package, a tool for transforming the Ecological Metadata Language (EML) in 
+JavaScript Object Notation, Linked Data (JSON-LD) format and subsequently working with these metadata files as semantically
+linked data, following the Resource Description Framework (RDF).  This helps mobilize the hundreds of thousands of 
+EML metadata files that have been and continue to be generated by major ecological research efforts to interface with
+RDF-based tools favored by information researchers and JSON-based tools favored by many developers. 
+
+
+## Test environments
+* local OS X install, R 3.5.2
+* ubuntu 14.04 (on travis-ci), R 3.5.2
+* win-builder (devel and release)
+
+## R CMD check results
+
+0 errors | 0 warnings | 1 note
+
+* This is a new release.
